fix: Race condition with cgroup
https://docs.kernel.org/admin-guide/cgroup-v2.html: "cgroups form a tree structure
and every process in the system belongs to one and only one cgroup. All threads of
a process belong to the same cgroup. On creation, all processes are put in the
cgroup that the parent process belongs to at the time. A process can be migrated
to another cgroup. **Migration of a process doesn’t affect already existing descendant
processes**"

D-Bus method calls are executed asynchronously. To block command execution until the
cgroup is created, it is necessary to explicitly wait for the job completion. Otherwise,
the child process will execute the exec system call earlier than necessary. This means
the child process can fork before placing itself in the cgroup. From the paragraph above,
it is clear that such child processes will no longer be in the required cgroup.

I spent the whole day on this. I hate multithreaded programming...